requirements.txt
fastapi~=0.103.1
pydantic~=2.3.0
sqlalchemy
alembic
psycopg2-binary
python-dotenv
httpx
pytest